move view_space_routes to own file + tests
